refactor(midi): share the bounded packet decoder
